requirements.txt
discord.py[voice]
python-dotenv
asyncio
requests